requirements.txt
kivy
https://github.com/kivymd/kivymd/archive/master.zip
pyjnius
materialyoucolor
pillow
asynckivy
asyncgui